src/main/java/com/gtx/sell/service/impl/SellerServiceImpl.java
package com.gtx.sell.service.impl;

import com.gtx.sell.dao.SellerInfo;
import com.gtx.sell.repository.SellerInfoRepository;
import com.gtx.sell.service.SellerService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
@Slf4j
public class SellerServiceImpl implements SellerService {

    @Autowired
    private SellerInfoRepository sellerInfoRepository;

    @Override
    public SellerInfo findSellerInfoByOpenid(String openid) {
        return sellerInfoRepository.findByOpenid(openid);
    }
}
